Refactored and clean up push notifcation setting up code.
Added 2 static methods for Playbasis to handle the native code usually lies in Delegate file class. As of now, we don't make Playbasis as a singleton clas, but instead using public static method.

sDeviceRetrievalKey is NSString* that Playbasis sdk will use later on in the process of calling registerForPushNotification(). It will use this key to retrieve processed device-token saved in NSUserDefaults.