Add a (still manual) queue to dispatch http request when needed.
In turns, this commit separates functionalities into different class file. But the main inclusion (header include) and the way to use it in code are still the same via 'playbasis.h'.

This commit have a button. An HTTP request will be created and enqueued to a queue. Then manually touch on that button to dispatch it. There's a work needed to be done to make the process automatic ie. looping every interval period to dispatch requests in a queue one by one.