Wire-up save/load at the proper stage during application cycle.
Added listening to UIApplication's events to hide all internal code from users in order to make saving/loading of request operational queue works as expected.

save() will serialize and save all http requests (PBRequest) into a file (default at 'requests.pb'. Then it will remove all objects residing in the queue for consistency.

load() will only do the work if the queue is in fresh state or it will exit the app immediately (for now). It will try to see if the file 'requests.pb' is there. If it is, then continue loading, and finally remove the file for consistency. Otherwise, if it's not then it means the queue would start fresh this time.