Refactor playerPublic json-response and PBRequest to response to specific type of request.
It's only for playerPublic: API call right now. Refactoring for specific response type has a challange to deal with. Block cannot be type-checked dynamically, and our base code store delegate as basic type of id<PBResponseHandler> whicn in turn it cannot be efficiently in method-conform checking in dynamic (we need to do casting everytime which is not good).

Thus we use add enum 'pbResponseType' to distingquish each type of response.

processResponse... for all response will response only information request if success, otherwise response only error (NSError) if the request fails, or there's error code returned from Playbasis platform. This means if error happens as it returned error code from Playbasis platform in the message, we will create NSError* object and send them into response method. This is a clean way, and flexible for user to use information upon response.