Add refactored json-response support for levels: API enI end-point.
'null' can be returned from server from the last level. Thus we check this whether before the converting process, it can responds to @selector(unsignedIntegerValue) or not. If not, then we set its value to maximum value of unsigned integer. It's not nice, but we need to assign a value to it, not null.