Fix @property syntax in PBRequest. Fix async url request to have a propert response back.
This changes has the base code for async url request that can be used by other api calls. PBRequest now receives syncUrl (BOOL) as it will check whenever it receive the response back whether which response type should be send to.

PBResultStatus_Response is created to apply for those api calls that only need a result status with no payload.

Added only login: for example and as a base code for async url request for this change.