Refactor code for method naming of mission:
This kind of method calling is straight forward in its own sense, thus we should not inject their name that much. This is to preserve the cleanness. Thus for mission:, we rearranged a first two parameters. So player:, or even mission:, the first parameter should be involving around the method name thus player-id, or mission-id.