Add all possible optional parameters for pointHistory: api call.
Changed its name to pointHistoryOfPlayer: as well.

We will adhere to 'Of' as part of the method name whenever its operation is to get information for such player.

Another of 'For', we will use this as part of method name whenever we do an action ie. delete, answer quiz, for such player.