Add json-response support actionLastPerformedForPlayer:
This added json-respons support for this type of response type applies a clean way by separating response data class from its wrapper response class.

The reason behind separation is that we need data class to be declare and use across the project in which it refers to the data directly. Wrapper response class just wrap the data, its class name is not fully suitable to be refered to across project in practical normal usage of users. But its wrapper class can be added more information (ie. properties such as processing time, version, timestamp etc.) later on if needed in the future.