Add encryption in building phase for api-key and api-secret. Add a couple more set for auth, and renew for more secure way to authenticate the app.
pblib (the name of the project in XCode for Playbasis library) has its own bundle. To reduce the exposure level of api-keys configuration of the client app, we then let users define it in configuration file which is not included in the app bundle, but it will be included later after it gets encrypted during the build phase.

We bundle 'crypt' app to encrypt the input file and write out to output file path. This program works alongside pblib in building phase.