Change the way in caching.
It's not just tying down to mainmenu. We created a user's globalCaching (it's not part of SDK) to cache specific data. After user make a request to get newer data back, then it update to globalCaching as well. Thus next time when the network cannot be reachable, then the newest data is available.

This applied to quest demo.

Also refactored much of code in questPageViewController class.