Set properly PBRequestUnit's state during processing.
Initially it's not that much useful as for block response we didn't send a request instance to it so there's no way to get state or peek into other information exposed by such a request. Also, there's very less likely to use a request instance.

Anyway, we set properly state from the start til finish.