Fix bug of Auth returning from error too early.
Auth API call consists of 2 hop before returning the success result as of before. This will fix a bug of returning an error from 1st hop without notifying user's delegate, so it will result in no more progress in the flow.

This will fix it, and allow user's flow to continue and be able to retry if needed.