Add pre-fetched OCMapper 2.1 as custom podspec
We need to do pre-fetching as OCMapper's podspec file originally allow for iOS platform only.
Thus we add deployment target support for OSX into podspec file, and it then successfully fetch OCMapper to use with the whole project.

OCMapper is configured to use with Playbasis, and PlaybasisMacOS targets. The code is not integrated yet, but external test projects test resulting frameworks (both Playbasis.framework, and OCMapper.framework, and also for Mac OS) successfully.