Add more quiz api
There might be a bug in OCMapper that causes JSON field not to be deserialized in which its root cause might be because of inheritance of a class that has "desc" (mapped from "description"). This makes derived class needs to manually map it as well to avoid crash. Contacted OCMapper's author already.